splits up the various classes into separate files
